requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
chromedriver==2.24.1
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
idna==2.9
numpy==1.18.5
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
robotframework==3.2.1
robotframework-seleniumlibrary==4.4.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
webdriver-manager==3.2.1
